jumpinng off walls work now
it turns out my old logic of jumping was wrong -- I did not let ppl jump if the player is already in the air, but if player was top of a wall, that's in the air so no jumpig!  so now we allow jumping off the top of the wall.

I also added different colours for walls so G for grass wall, M for mud wall, and W is still black wall.