walking on walls now work!
it turns out in Python you cannot just assign a rectangle to a "new" rectangle.. that is just like giving a new nickname to the rectangle.. you have to really make a new one.  So futurebox was just pointing to self.box.. now it works after this fix and other fixes to the way we check walls and boundary.. also now we check walls first before boundary .. becuase walls are inside boundaries always!